backend/requirements.txt
fastapi
uvicorn
motor
boto3
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
whisper
pyannote.audio
celery
redis
python-dotenv